requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
faiss-cpu>=1.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
